Update pinning to resolve dependabot warning about vulnerable Django
Quote:
In Django 3.2 before 3.2.17, 4.0 before 4.0.9, and 4.1 before 4.1.6, the parsed values of Accept-Language headers are cached in order to avoid repetitive parsing. This leads to a potential denial-of-service vector via excessive memory usage if the raw value of Accept-Language headers is very large.

See: https://github.com/Klimaat-Helpdesk/wagtail-helpdesk/security/dependabot/2